requirements.txt
asgiref==3.3.4
autopep8==1.5.6
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
devtools==0.6.1
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
flake8==3.9.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Janome==0.4.1
livereload==2.6.3
mccabe==0.6.1
phonenumbers==8.12.22
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
rfc3986==1.5.0
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
urllib3==1.26.4
webdriver-manager==3.4.1
whitenoise==5.2.0
gunicorn==20.0.4